+++ templates/artworks.html
@@ -94,6 +94,6 @@
     </div>
   </div>
 
-  <script src="/static/script.js"></script>
+  <script type="module" src="/static/modules/main.js"></script>
 </body>
 </html>
