+++ templates/quiz.html
@@ -26,6 +26,6 @@
       </div>
     </section>
   </div>
-  <script src="/static/script.js"></script>
+  <script type="module" src="/static/modules/main.js"></script>
 </body>
 </html>
